fixed TTL feedback
